test: move focused cleanup fixture to conftest
